app/requirements.txt
Flask
mysql-connector
pycrypto
cryptography
pprp
requests
uwsgi